+++ Source/LibreArp.cpp
@@ -19,21 +19,24 @@
 #include "editor/MainEditor.h"
 #include "util/MathConsts.h"
 
-const juce::Identifier LibreArp::TREEID_LIBREARP = "libreArpPlugin"; // NOLINT
-const juce::Identifier LibreArp::TREEID_LOOP_RESET = "loopReset"; // NOLINT
-const juce::Identifier LibreArp::TREEID_PATTERN_XML = "patternXml"; // NOLINT
-const juce::Identifier LibreArp::TREEID_OCTAVES = "octaves"; // NOLINT
-const juce::Identifier LibreArp::TREEID_SMART_OCTAVES = "smartOctaves"; // NOLINT
-const juce::Identifier LibreArp::TREEID_INPUT_VELOCITY = "usingInputVelocity"; // NOLINT
-const juce::Identifier LibreArp::TREEID_SWING = "swing"; // NOLINT
-const juce::Identifier LibreArp::TREEID_MAX_CHORD_SIZE = "maxChordSize"; // NOLINT
+const juce::Identifier LibreArp::TREEID_LIBREARP                   = "libreArpPlugin"; // NOLINT
+const juce::Identifier LibreArp::TREEID_LOOP_RESET                 = "loopReset"; // NOLINT
+const juce::Identifier LibreArp::TREEID_PATTERN_XML                = "patternXml"; // NOLINT
+const juce::Identifier LibreArp::TREEID_OCTAVES                    = "octaves"; // NOLINT
+const juce::Identifier LibreArp::TREEID_SMART_OCTAVES              = "smartOctaves"; // NOLINT
+const juce::Identifier LibreArp::TREEID_INPUT_VELOCITY             = "usingInputVelocity"; // NOLINT
+const juce::Identifier LibreArp::TREEID_SWING                      = "swing"; // NOLINT
+const juce::Identifier LibreArp::TREEID_MAX_CHORD_SIZE             = "maxChordSize"; // NOLINT
 const juce::Identifier LibreArp::TREEID_EXTRA_NOTES_SELECTION_MODE = "extraNotesSelectionMode"; // NOLINT
-const juce::Identifier LibreArp::TREEID_NUM_INPUT_NOTES = "numInputNotes"; // NOLINT
-const juce::Identifier LibreArp::TREEID_OUTPUT_MIDI_CHANNEL = "outputMidiChannel"; // NOLINT
-const juce::Identifier LibreArp::TREEID_INPUT_MIDI_CHANNEL = "inputMidiChannel"; // NOLINT
-const juce::Identifier LibreArp::TREEID_NON_PLAYING_MODE_OVERRIDE = "nonPlayingModeOverride"; // NOLINT
-const juce::Identifier LibreArp::TREEID_BYPASS = "bypass"; // NOLINT
-const juce::Identifier LibreArp::TREEID_PATTERN_OFFSET = "patternOffset"; // NOLINT
+const juce::Identifier LibreArp::TREEID_NUM_INPUT_NOTES            = "numInputNotes"; // NOLINT
+const juce::Identifier LibreArp::TREEID_OUTPUT_MIDI_CHANNEL        = "outputMidiChannel"; // NOLINT
+const juce::Identifier LibreArp::TREEID_INPUT_MIDI_CHANNEL         = "inputMidiChannel"; // NOLINT
+const juce::Identifier LibreArp::TREEID_NON_PLAYING_MODE_OVERRIDE  = "nonPlayingModeOverride"; // NOLINT
+const juce::Identifier LibreArp::TREEID_BYPASS                     = "bypass"; // NOLINT
+const juce::Identifier LibreArp::TREEID_PATTERN_OFFSET             = "patternOffset"; // NOLINT
+const juce::Identifier LibreArp::TREEID_USER_TIME_SIG              = "userTimeSig"; // NOLINT
+const juce::Identifier LibreArp::TREEID_USER_TIME_SIG_NUMERATOR    = "userTimeSigNumerator"; // NOLINT
+const juce::Identifier LibreArp::TREEID_USER_TIME_SIG_DENOMINATOR  = "userTimeSigDenominator"; // NOLINT
 
 static const int NOTES_IN_OCTAVE = 12;
 
@@ -238,8 +241,8 @@ void LibreArp::processMidi(int numSamples, juce::MidiBuffer& midi) {
         fillCurrentNonPlayingPositionInfo(cpi);
     }
 
-    this->timeSigNumerator = cpi.timeSigNumerator;
-    this->timeSigDenominator = cpi.timeSigDenominator;
+    this->hostTimeSigNumerator = cpi.timeSigNumerator;
+    this->hostTimeSigDenominator = cpi.timeSigDenominator;
 
     // Output generation
     if (!*bypass && cpi.isPlaying && !this->events.events.empty() && this->events.loopLength > 0) {
@@ -446,6 +449,15 @@ void LibreArp::setStateInformation(const void *data, int sizeInBytes) {
             if (tree.hasProperty(TREEID_PATTERN_OFFSET)) {
                 this->patternOffset = static_cast<juce::int64>(tree.getProperty(TREEID_PATTERN_OFFSET));
             }
+            if (tree.hasProperty(TREEID_USER_TIME_SIG)) {
+                this->userTimeSig = tree.getProperty(TREEID_USER_TIME_SIG);
+            }
+            if (tree.hasProperty(TREEID_USER_TIME_SIG_NUMERATOR)) {
+                this->userTimeSigNumerator = tree.getProperty(TREEID_USER_TIME_SIG_NUMERATOR);
+            }
+            if (tree.hasProperty(TREEID_USER_TIME_SIG_DENOMINATOR)) {
+                this->userTimeSigDenominator = tree.getProperty(TREEID_USER_TIME_SIG_DENOMINATOR);
+            }
 
             setPattern(loadedPattern);
         }
@@ -470,6 +482,9 @@ juce::ValueTree LibreArp::toValueTree() {
     tree.setProperty(TREEID_NON_PLAYING_MODE_OVERRIDE, NonPlayingMode::toJuceString(this->nonPlayingModeOverride), nullptr);
     tree.setProperty(TREEID_BYPASS, this->bypass->get(), nullptr);
     tree.setProperty(TREEID_PATTERN_OFFSET, static_cast<juce::int64>(this->patternOffset), nullptr);
+    tree.setProperty(TREEID_USER_TIME_SIG, this->userTimeSig, nullptr);
+    tree.setProperty(TREEID_USER_TIME_SIG_NUMERATOR, this->userTimeSigNumerator, nullptr);
+    tree.setProperty(TREEID_USER_TIME_SIG_DENOMINATOR, this->userTimeSigDenominator, nullptr);
     return tree;
 }
 
@@ -535,12 +550,42 @@ int LibreArp::getNumInputNotes() const {
     return this->octaveSize;
 }
 
+
+void LibreArp::setUserTimeSig(bool v) {
+    this->userTimeSig = v;
+}
+
+bool LibreArp::isUserTimeSig() const {
+    return this->userTimeSig;
+}
+
+void LibreArp::setUserTimeSigNumerator(int v) {
+    this->userTimeSigNumerator = v;
+}
+
+int LibreArp::getUserTimeSigNumerator() const {
+    return this->userTimeSigNumerator;
+}
+
+void LibreArp::setUserTimeSigDenominator(int v) {
+    this->userTimeSigDenominator = v;
+}
+
+int LibreArp::getUserTimeSigDenominator() const {
+    return this->userTimeSigDenominator;
+}
+
+
 int LibreArp::getTimeSigNumerator() const {
-    return this->timeSigNumerator;
+    return (this->userTimeSig)
+        ? this->userTimeSigNumerator
+        : this->hostTimeSigNumerator;
 }
 
 int LibreArp::getTimeSigDenominator() const {
-    return this->timeSigDenominator;
+    return (this->userTimeSig)
+        ? this->userTimeSigDenominator
+        : this->hostTimeSigDenominator;
 }
 
 void LibreArp::fillCurrentNonPlayingPositionInfo(juce::AudioPlayHead::CurrentPositionInfo &cpi) {